feat: 模版的获取从资产包中来
