+++ packages/gi-assets-basic/src/templates/index.tsx
@@ -1,0 +1,3 @@
+import Query from './Query';
+import Simple from './Simple';
+export { Simple, Query };
