+++ packages/gi-site/src/services/template.ts
@@ -1,13 +1,10 @@
 import { message } from 'antd';
-
 import { GI_TEMPLATE_DB } from '../hooks/useUpdate';
 import { getUid } from '../pages/Workspace/utils';
+import { queryAssets } from './assets';
 import { GI_SITE } from './const';
-import { TEMPLATE_QUERY } from './initial.data/query.template';
-import { TEMPLATE_SIMPLE } from './initial.data/simple.template';
 import { ITemplate } from './typing';
 import { request } from './utils';
-
 /**
  * 获取所有项目
  * @returns
@@ -16,13 +13,13 @@ export const list = async (type: 'my' | 'graph'): Promise<ITemplate[]> => {
   if (GI_SITE.IS_OFFLINE) {
     if (type === 'graph') {
       //@ts-ignore
-      return [TEMPLATE_SIMPLE, TEMPLATE_QUERY];
+      const { templates } = await queryAssets();
+      return Object.values(templates);
     } else {
       const res: ITemplate[] = [];
       await GI_TEMPLATE_DB.iterate((item: ITemplate) => {
         res.push(item);
       });
-      console.log('RES', res);
 
       return res;
     }
